+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project are recorded here.
 The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+[Unreleased]
+
+TODO
+
+## [0.3.0] - 2026-08-15
 
 ### Added
 
@@ -147,6 +151,7 @@ can make a whole changeset.
 
 Initial release.
 
-[Unreleased]: https://github.com/jonasdeyson/intact/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/jonasdeyson/intact/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/jonasdeyson/intact/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/jonasdeyson/intact/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/jonasdeyson/intact/releases/tag/v0.1.0